fix: use web standard event apis for twilio websocket (#127)
* fix: use web standard event apis for twilio websocket

the twilio websocket extension used the node `ws` specific apis for attaching event handlers, which made it throw in a cloudflare environment. this patch adds a fix to the types, and uses `.addEventListener` on the websocket instead of `.on`. this also adds a fix to load the right shim for workerd when using `@openai/agents-realtime`

* add a changeset

* Update shiny-berries-kiss.md

* pass tests

* pass test

* use websocket protocols in workerd when making a new websocket connection